+++ elacca/plugin-outputs/server-index.tsx
@@ -1,7 +1,0 @@
-"skip ssr";
-
-class Page extends React.Component {}
-function DefaultExportRenamedByElacca() {
-  return null;
-}
-export default DefaultExportRenamedByElacca;