+++ example-app/plugin-outputs/server-index.tsx
@@ -1,26 +1,9 @@
 'skip ssr';
 
-import _default from "react";
-import { Checkbox, Select } from '@chakra-ui/react';
+import { Checkbox, FormControl, FormLabel, HStack, PinInput, PinInputField, Select, Slider, SliderFilledTrack, SliderThumb, SliderTrack, Switch, Textarea } from '@chakra-ui/react';
 import { ChakraProvider } from '@chakra-ui/react';
 import React from 'react';
-function HeavyComponent({
-  hello
-}) {
-  return <div className='w-[600px]'>
-            <div className=''>
-                <Select placeholder='Select option'>
-                    <option value='option1'>Option 1</option>
-                    <option value='option2'>Option 2</option>
-                    <option value='option3'>Option 3</option>
-                </Select>
-            </div>
-            <div className=''>
-                <Checkbox defaultChecked>Checkbox</Checkbox>
-            </div>
-            <h1>Heavy Component {hello}</h1>
-        </div>;
-}
+import { createContext } from 'react';
 export function getServerSideProps() {
   return {
     props: {
@@ -28,6 +11,7 @@ export function getServerSideProps() {
     }
   };
 }
+const context = createContext({});
 function DefaultExportRenamedByElacca() {
   return null;
 }