+++ example-app/src/pages/index.tsx
@@ -16,6 +16,7 @@ import {
 } from '@chakra-ui/react'
 import { ChakraProvider } from '@chakra-ui/react'
 import React from 'react'
+import { createContext } from 'react'
 
 export default function HeavyComponent({ hello }) {
     return (
@@ -65,3 +66,5 @@ export function getServerSideProps() {
         },
     }
 }
+
+const context = createContext({})