+++ elacca/CHANGELOG.md
@@ -1,5 +1,12 @@
 # elacca
 
+## 0.0.2
+
+### Patch Changes
+
+-   remove unused functions when in server
+-   annotate functions as pure
+
 ## 0.0.1
 
 ### Patch Changes