+++ example-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # nextjs-app
 
+## 0.1.2
+
+### Patch Changes
+
+-   Updated dependencies
+-   Updated dependencies
+    -   elacca@0.0.2
+
 ## 0.1.1
 
 ### Patch Changes